+++ src/main/java/keeper/project/homepage/admin/service/library/BookManageService.java
@@ -1,12 +1,10 @@
 package keeper.project.homepage.admin.service.library;
 
-import java.awt.print.Book;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
-import java.util.Optional;
-import keeper.project.homepage.dto.library.BookDto;
+import keeper.project.homepage.admin.dto.library.BookDto;
 import keeper.project.homepage.dto.result.CommonResult;
 import keeper.project.homepage.entity.ThumbnailEntity;
 import keeper.project.homepage.entity.library.BookBorrowEntity;
