+++ src/main/java/keeper/project/homepage/user/controller/library/LibraryMainController.java
@@ -1,10 +1,8 @@
 package keeper.project.homepage.user.controller.library;
 
-import keeper.project.homepage.dto.library.BookDto;
-import keeper.project.homepage.dto.library.BookResult;
+import keeper.project.homepage.user.dto.library.BookResult;
 import keeper.project.homepage.dto.result.ListResult;
 import keeper.project.homepage.dto.result.SingleResult;
-import keeper.project.homepage.entity.library.BookEntity;
 import keeper.project.homepage.service.ResponseService;
 import keeper.project.homepage.user.service.library.LibraryMainService;
 import lombok.RequiredArgsConstructor;
